Adjust formatting

+++ features/merchant/shop_discount/src/main/java/com/tokopedia/shopdiscount/manage/data/response/GetSlashPriceProductListResponse.kt
@@ -29,7 +29,7 @@ data class GetSlashPriceProductListResponse(
             @SerializedName("discounted_percentage")
             val discountedPercentage: Int = 0,
             @SerializedName("discounted_price")
-            val discountedPrice: Int = 0,
+            val discountedPrice: Long = 0,
             @SerializedName("discounted_price_data")
             val discountedPriceData: DiscountedPriceData = DiscountedPriceData(),
             @SerializedName("end_date")
